parser/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
marker-pdf==0.2.17
sentence-transformers==3.1.1
torch==2.4.1
